pyproject.toml
[tool.poetry]
name = "repl_python3_Rancher"
version = "0.1.0"
description = ""
authors = ["HunterBerg <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
ffmpeg = "^1.4"
python-dotenv = "^0.17.1"
youtube-dl = "^2021.4.26"
PyNaCl = "^1.4.0"
flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
